feat: `exists` and `id` API endpoints
The `/exists` is mainly useful for user registration when the client
needs to validate whether a particular provided username is already in
use.

The `/id` endpoint can be used to retrieve the ID associated with a
username. Mainly used for the login system where an ID is required to
make `/auth` requests.